Adding interaction terms in LPL model (#258)
As per discussion, the model was fitted given the data with a new column
including season-state combo from 2009-08-01 to 2021-04-01 and was
evaluated.

Overall, the fit was good. No divergence of chains (R_hat is within
[1.0,1.01]). The posterior density plot and the trace plot of parameters
for interaction terms were checked. The posterior looked fine, and the
chains looked well-mixed:

<img width="950" height="218" alt="image"
src="https://github.com/user-attachments/assets/7f8f36f9-8243-4a34-9eea-357066e39757"
/>

Similar for M:
<img width="950" height="218" alt="image"
src="https://github.com/user-attachments/assets/a5da28f4-3dd7-4c9a-a2d9-31d38d7b16a5"
/>

Then the in-sample posterior predictive distribution was checked. The
prediction mean as well as the interval was plotted against the data for
all the seasons and states:
<img width="4323" height="16750" alt="fit_inter"
src="https://github.com/user-attachments/assets/4954f092-6276-4f38-9f8f-349b5c5a39d8"
/>

Sorry for the humungo picture, but this gives us an overview of how well
each combo is fitted! Overall, most of the prediction mean (or at least
the prediction interval) overlapped with the confidence interval of the
data.

Comparing to the previous fit without interaction term: 
<img width="4323" height="16850" alt="fit"
src="https://github.com/user-attachments/assets/0834caf4-9141-4b99-bf6a-727aff9a8f59"
/>

The fitting in 2021/2022 season for most of the states has improved.
Some combos like "Missouri-21/22", "Alabama-21/22", and most of the
seasons in District of Columbia were improved.

I think this is great sign to continue on the track of adding
interaction term! What are your thoughts? @swo @afmagee42

---------

+++ iup/utils.py
@@ -109,9 +109,7 @@ def value_to_index(
         ) - set(mapping[col_name].keys()):
             raise RuntimeError(f"Missing indices for values: {missing_values}")
 
-        groups = groups.with_columns(
-            pl.col(col_name).replace_strict(mapping[col_name]).cast(pl.UInt8)
-        )
+        groups = groups.with_columns(pl.col(col_name).replace_strict(mapping[col_name]))
 
     array = groups.to_numpy() + np.cumsum([0] + num_group_levels[:-1])
